implement get_type, which returns the supposed type for an AST node

+++ tests/foo.c
@@ -1,4 +1,3 @@
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -12,9 +11,9 @@ int fooxy(int x, int y) {
     return 0;
 }
 
-int alloc4(int64_t **p, int64_t a, int64_t b, int64_t c, int64_t d) {
+int alloc4(int **p, int a, int b, int c, int d) {
     *p = malloc(4 * sizeof(int));
-    int64_t *q = *p;
+    int *q = *p;
     q[0] = a;
     q[1] = b;
     q[2] = c;
